mise a jour avec l'ia rapport et autre suijet important

+++ backend/src/main/java/com/lexsecura/application/dto/AuditVerifyResponse.java
@@ -2,7 +2,7 @@ package com.lexsecura.application.dto;
 
 public record AuditVerifyResponse(
         boolean valid,
-        int totalEvents,
+        long totalEvents,
         int verifiedEvents,
         String message
 ) {}
